+++ Changelog.md
@@ -1,5 +1,13 @@
 # Version History
 
+- 0.4.0
+  
+  - Fixed Vulkan Swapchain on resize occasionally result in error messages
+  - Revisited Examples and made them use SDL3 for window creation
+  - Added common library for examples
+  - Added Multiview support with example
+  - Examples are now stored in examples/bin folder with compiled shaders and resources
+
 - 0.3.15
 
   - Fixed Vulkan default texture barriers